refactor(freemarker): 重構 FreeMarker 方法，更新包名並移除不必要的類別
- 將 ApiResult 的引用從 pojo 更新為 dto。
- 刪除不再使用的 TrimMethod 和 EscapeXmlMethod 類別。
- 新增自定義的 TrimMethod 類別，處理字串的前後空白及 null 值。
- 新增 MarkdownToHtml 類別，實現 Markdown 轉 HTML 的功能。
- 更新 FtlRenderer 類別的文檔註釋，提供更清晰的說明。
