+++ src/main/java/com/vance/demo/controller/TestController.java
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.google.common.collect.Maps;
+import com.vance.demo.dto.ApiResult;
 import com.vance.demo.model.User;
-import com.vance.demo.pojo.ApiResult;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
