+++ src/main/java/com/vance/demo/util/common/StringUtil.java
@@ -1,45 +1,17 @@
-package com.vance.demo.support;
+package com.vance.demo.util.common;
 
 import java.math.BigDecimal;
-import java.util.Collection;
-import java.util.Map;
 
-import org.apache.commons.collections4.CollectionUtils;
-import org.apache.commons.collections4.MapUtils;
+import org.apache.commons.lang3.StringUtils;
 
 import com.vance.demo.constant.Constant;
 
 /**
- * 工具類別
+ * 字串處理工具
  * 
  * @author Vance
  */
-public class Util {
-    /**
-     * 當傳入值為null時，回傳一個不可異動的空物件
-     *
-     * @param <T>      the element type
-     * @param iterable the iterable(List, Set)
-     * @return
-     * @since Apache commons-collections4 CollectionUtils
-     */
-    public static <T> Iterable<T> emptyIfNull(final Collection<T> collection) {
-        return CollectionUtils.emptyIfNull(collection);
-    }
-
-    /**
-     * 當傳入值為null時，回傳一個不可異動的空物件
-     *
-     * @param <K> the key type
-     * @param <V> the value type
-     * @param map the map, possibly <code>null</code>
-     * @return an empty map if the argument is <code>null</code>
-     * @since Apache commons-collections4 MapUtils
-     */
-    public static <K, V> Map<K, V> emptyIfNull(final Map<K, V> map) {
-        return MapUtils.emptyIfNull(map);
-    }
-
+public class StringUtil {
     /**
      * 去除前後的空白字元。
      *
@@ -48,7 +20,7 @@ public class Util {
      */
     public static String trim(Object str) {
         if (str == null) {
-            return "";
+            return StringUtils.EMPTY;
         } else if (str instanceof Double) {
             BigDecimal bd = BigDecimal.valueOf((Double) str);
             return bd.toPlainString();
@@ -87,16 +59,5 @@ public class Util {
             sb.append("0");
         sb.append(str);
         return sb.toString();
-    }
-
-    /**
-     * 是否符合freemarker exception string格式 (因為當傳入的字串為null會變成
-     * freemarker.core.DefaultToExpression$EmptyStringAndSequenceAndHash)
-     * 
-     * @param value
-     * @return
-     */
-    public static boolean isFreemarkerEmptyString(Object value) {
-        return trim(value).toUpperCase().matches("^FREEMARKER\\.CORE\\.DEFAULTTOEXPRESSION.*$");
     }
 }
